A new devtest world generates around the player and is drawn
Opening a world whose mg_name is v7 now puts the player on generated
ground with the world drawn around them, which is the first item of the
master plan's list. What was in the way:

The noise hash multiplied in signed arithmetic, where it is meant to
overflow. An optimiser that takes the absence of signed overflow at its
word gives back values outside -1...1, and every octave came out biased
about one high: a terrain noise of offset 4 and scale 70 answered 200
everywhere, so v7 was solid rock to the top of the world and infidigger's
own terrain was a good deal fuller than it was written to be. Luanti does
the same arithmetic in unsigned, and so does this now. check_map's noise
check takes 400 samples of one octave and fails if they leave -1...1 or
average more than 0.15 away from zero.

The vendored mapgens write air by its content id rather than by name, so
CONTENT_AIR, CONTENT_IGNORE and CONTENT_UNKNOWN in the shim's mapnode.h
are the ids the module reserves rather than Luanti's 125, 126 and 127 --
a v7 world was written with a testnodes rail where its air should be.
VendoredGenerator's constructor checks the three against what the module
sent. The node definition shim answers "found" for a name it knows rather
than "not CONTENT_IGNORE", since ignore is now id zero and the default
biome names it as its cave liquid. And the EmergeParams a mapgen is built
with belongs to the mapgen, which deletes it: it is allocated for that
now, it owns the managers and the biome generator the way Luanti's does,
and a v7 generator built and thrown away at startup is the check that it
adds up.

check_map works above where any mapgen builds, in sections the module
pins while it runs -- a generator writes only air up there, and nothing
else keeps those sections loaded. The claim that an unbuilt voxel reads
as air is a singlenode claim and is made only in a singlenode world.

A player who has never been here stands on the ground: core.get_spawn_level
reads the ground out of a column of the map, and a player who arrives
before there is any ground to read is stood on it as soon as there is,
unless they have moved by then. And the launcher's menu stays closed once
the world is up -- the save list it asked for arrives after the world,
because the server was busy loading the game, and drew itself over it.

+++ src/impl/noise.cpp
@@ -55,21 +55,27 @@ int PseudoRandom::range(int min, int max)
 
 
 //noise poly:  p(n) = 60493n^3 + 19990303n + 137612589
+// The hash is unsigned because it is meant to overflow: signed overflow is
+// undefined, and an optimiser that takes it at its word gives back values
+// outside -1...1, which is a mapgen's terrain biased upwards by an octave.
+// Luanti's own does the same arithmetic in unsigned and casts to signed for
+// the division, so this stays the same noise as the one a Luanti world was
+// generated with.
 float noise2d(int x, int y, int seed) {
-	int n = (NOISE_MAGIC_X * x + NOISE_MAGIC_Y * y
+	unsigned int n = (NOISE_MAGIC_X * x + NOISE_MAGIC_Y * y
 			+ NOISE_MAGIC_SEED * seed) & 0x7fffffff;
 	n = (n >> 13) ^ n;
 	n = (n * (n * n * 60493 + 19990303) + 1376312589) & 0x7fffffff;
-	return 1.f - (float)n / 0x40000000;
+	return 1.f - (float)(int)n / 0x40000000;
 }
 
 
 float noise3d(int x, int y, int z, int seed) {
-	int n = (NOISE_MAGIC_X * x + NOISE_MAGIC_Y * y + NOISE_MAGIC_Z * z
+	unsigned int n = (NOISE_MAGIC_X * x + NOISE_MAGIC_Y * y + NOISE_MAGIC_Z * z
 			+ NOISE_MAGIC_SEED * seed) & 0x7fffffff;
 	n = (n >> 13) ^ n;
 	n = (n * (n * n * 60493 + 19990303) + 1376312589) & 0x7fffffff;
-	return 1.f - (float)n / 0x40000000;
+	return 1.f - (float)(int)n / 0x40000000;
 }
